:white_check_mark::construction: Moved ShipmentRequest tests to own test file.
