+++ tests/Feature/ShipmentRequestTest.php
@@ -2,23 +2,17 @@
 
 declare(strict_types=1);
 
-namespace MyParcelCom\Microservice\Tests\Endpoints;
+namespace MyParcelCom\Microservice\Tests\Feature;
 
 use Mockery;
 use MyParcelCom\Hermes\Http\ShipmentRequest;
 use MyParcelCom\Microservice\Tests\Mocks\ShipmentRequestMock;
 use MyParcelCom\Microservice\Tests\TestCase;
 use MyParcelCom\Microservice\Tests\Traits\CommunicatesWithCarrier;
-use MyParcelCom\Microservice\Tests\Traits\JsonApiAssertionsTrait;
 
-/**
- * @group Endpoints:Shipment
- * @group Implementation
- */
-class ShipmentsTest extends TestCase
+class ShipmentRequestTest extends TestCase
 {
     use CommunicatesWithCarrier;
-    use JsonApiAssertionsTrait;
 
     protected function tearDown()
     {
@@ -27,25 +21,23 @@ class ShipmentsTest extends TestCase
     }
 
     /** @test */
-    public function testPostShipment()
+    public function testItValidatesAShipmentRequestBasedOnValidationRules()
     {
+        $this->app->singleton(ShipmentRequest::class, ShipmentRequestMock::class);
         $this->bindCarrierApiGatewayMock();
-
-        // TODO: Add carrier response stub for creating a shipment.
-        // See the "Response Stubs" chapter in the readme for more info.
 
         $data = \GuzzleHttp\json_decode(
             file_get_contents(base_path('tests/Stubs/shipment-request.stub')),
             true
         );
 
-        $this->assertJsonSchema(
-            '/shipments',
-            '/shipments',
-            $this->getRequestHeaders(),
-            $data,
-            'post',
-            201
-        );
+        $response = $this->json('post', '/shipments',$data, $this->getRequestHeaders());
+
+        dd($response->json());
+        // TODO: Add assertions!
     }
+
+    // TODO: Add test for when validation passes.
+
+    // TODO: Add tests for when validation fails for each case :) (see ShipmentRequestMock for the rules)
 }
